# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ELD Trip Planner API",
        "version": "1.0.0",
        "description": "Plans a truck trip and its FMCSA daily log sheets."
    },
    "paths": {
...
        "/api/v1/places/autocomplete": {
            "get": {
                "operationId": "autocompletePlaces",
                "description": "GET /api/v1/places/autocomplete?q=.",
                "summary": "US place suggestions",
                "parameters": [
                    {
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string",
                            "minLength": 3,
                            "maxLength": 100
                        },
                        "description": "3 to 100 characters after trimming. No control characters.",
                        "required": true
                    }
                ],
                "tags": [
                    "places"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AutocompleteResponse"
                                }
                            }
                        },
                        "description": "At most 5 suggestions. Cache-Control: public, max-age=3600."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "VALIDATION_ERROR: q is missing, too short, too long or has control characters."
                    },
                    "429": {
                        "headers": {
                            "Retry-After": {
                                "schema": {
                                    "type": "integer"
                                },
                                "description": "Seconds to wait, 1 to 300. Same value as error.retry_after_s."
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Rate limited. Read retry_after_s or the Retry-After header."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Unexpected error. Quote request_id."
                    },
                    "503": {
                        "headers": {
                            "Retry-After": {
                                "schema": {
                                    "type": "integer"
                                },
                                "description": "Seconds to wait, 1 to 300. Same value as error.retry_after_s."
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Routing service unavailable or out of quota (UPSTREAM_UNAVAILABLE, UPSTREAM_QUOTA_EXCEEDED)."
                    }
                }
            }
        },
        "/api/v1/trips/plan": {
            "post": {
                "operationId": "planTrip",
                "description": "POST /api/v1/trips/plan.",
                "summary": "Plan a trip with stops and daily log sheets",
                "tags": [
                    "trips"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlanTripRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanTripResponse"
                                }
                            }
                        },
                        "description": "The plan. Cache-Control: no-store."
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "VALIDATION_ERROR: the request must change. See field and details."
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "LOCATION_NOT_FOUND, UNSUPPORTED_LOCATION, ROUTE_NOT_FOUND or TRIP_TOO_LONG: the trip cannot be planned as entered."
                    },
                    "429": {
                        "headers": {
                            "Retry-After": {
                                "schema": {
                                    "type": "integer"
                                },
                                "description": "Seconds to wait, 1 to 300. Same value as error.retry_after_s."
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Rate limited. Read retry_after_s or the Retry-After header."
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Unexpected error. Quote request_id."
                    },
                    "503": {
                        "headers": {
                            "Retry-After": {
                                "schema": {
                                    "type": "integer"
                                },
                                "description": "Seconds to wait, 1 to 300. Same value as error.retry_after_s."
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Routing service unavailable or out of quota (UPSTREAM_UNAVAILABLE, UPSTREAM_QUOTA_EXCEEDED)."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AutocompleteResponse": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PlaceSuggestion"
                        }
                    }
                },
                "required": [
                    "items"
                ]
            },
...
            "ErrorBody": {
                "type": "object",
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                        "type": "string"
                    },
                    "field": {
                        "type": "string"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FieldError"
                        }
                    },
                    "retry_after_s": {
                        "type": "integer",
                        "maximum": 300,
                        "minimum": 1
                    },
                    "request_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "message",
                    "request_id"
                ]
            },
            "ErrorCode": {
                "enum": [
                    "VALIDATION_ERROR",
                    "LOCATION_NOT_FOUND",
                    "UNSUPPORTED_LOCATION",
                    "AMBIGUOUS_LOCATION",
                    "ROUTE_NOT_FOUND",
                    "TRIP_TOO_LONG",
                    "NOT_FOUND",
                    "METHOD_NOT_ALLOWED",
                    "RATE_LIMITED",
                    "UPSTREAM_QUOTA_EXCEEDED",
                    "UPSTREAM_UNAVAILABLE",
                    "INTERNAL"
                ],
                "type": "string",
                "description": "* `VALIDATION_ERROR` - VALIDATION_ERROR\n* `LOCATION_NOT_FOUND` - LOCATION_NOT_FOUND\n* `UNSUPPORTED_LOCATION` - UNSUPPORTED_LOCATION\n* `AMBIGUOUS_LOCATION` - AMBIGUOUS_LOCATION\n* `ROUTE_NOT_FOUND` - ROUTE_NOT_FOUND\n* `TRIP_TOO_LONG` - TRIP_TOO_LONG\n* `NOT_FOUND` - NOT_FOUND\n* `METHOD_NOT_ALLOWED` - METHOD_NOT_ALLOWED\n* `RATE_LIMITED` - RATE_LIMITED\n* `UPSTREAM_QUOTA_EXCEEDED` - UPSTREAM_QUOTA_EXCEEDED\n* `UPSTREAM_UNAVAILABLE` - UPSTREAM_UNAVAILABLE\n* `INTERNAL` - INTERNAL"
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "required": [
                    "error"
                ]
            },
            "FieldError": {
                "type": "object",
                "properties": {
                    "field": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "field",
                    "message"
                ]
            },
...
            "LogDay": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "description": "YYYY-MM-DD in the frozen trip offset."
                    },
                    "sheet_index": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "from_label": {
                        "type": "string"
                    },
                    "to_label": {
                        "type": "string"
                    },
                    "miles_driven": {
                        "type": "number",
                        "format": "double"
                    },
                    "segments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LogSegment"
                        }
                    },
                    "remarks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LogRemark"
                        }
                    },
                    "totals": {
                        "$ref": "#/components/schemas/LogTotals"
                    },
                    "recap": {
                        "$ref": "#/components/schemas/LogRecap"
                    }
                },
                "required": [
                    "date",
                    "from_label",
                    "miles_driven",
                    "recap",
                    "remarks",
                    "segments",
                    "sheet_index",
                    "to_label",
                    "totals"
                ]
            },
...
            "LogHeaderRequest": {
                "type": "object",
                "description": "Header text only; never affects HOS. Missing, null or blank all mean \"use the default\".",
                "properties": {
                    "driver_name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 80
                    },
                    "carrier_name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 100
                    },
                    "main_office_address": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 150
                    },
                    "home_terminal_address": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 150
                    },
                    "truck_number": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 40
                    },
                    "trailer_number": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 40
                    },
                    "shipping_doc": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 60
                    },
                    "shipper_commodity": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 100
                    }
                }
            },
...
            "PlaceSuggestion": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number",
                        "format": "double"
                    },
                    "lng": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "label",
                    "lat",
                    "lng"
                ]
            },
            "PlanTripRequest": {
                "type": "object",
                "properties": {
                    "current_location": {
                        "$ref": "#/components/schemas/LocationInput"
                    },
                    "pickup_location": {
                        "$ref": "#/components/schemas/LocationInput"
                    },
                    "dropoff_location": {
                        "$ref": "#/components/schemas/LocationInput"
                    },
                    "current_cycle_used_hours": {
                        "type": "number",
                        "format": "double",
                        "maximum": 70,
                        "minimum": 0
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "start_time": {
                        "type": "string",
                        "pattern": "^([01][0-9]|2[0-3]):(00|15|30|45)$"
                    },
                    "log_header": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/LogHeaderRequest"
                            }
                        ],
                        "nullable": true
                    }
                },
                "required": [
                    "current_cycle_used_hours",
                    "current_location",
                    "dropoff_location",
                    "pickup_location"
                ]
            },
            "PlanTripResponse": {
                "type": "object",
                "properties": {
                    "trip": {
                        "$ref": "#/components/schemas/TripMeta"
                    },
                    "route": {
                        "$ref": "#/components/schemas/Route"
                    },
                    "stops": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Stop"
                        }
                    },
                    "timeline": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TimelineEvent"
                        }
                    },
                    "days": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LogDay"
                        }
                    },
                    "summary": {
                        "$ref": "#/components/schemas/TripSummary"
                    }
                },
                "required": [
                    "days",
                    "route",
                    "stops",
                    "summary",
                    "timeline",
                    "trip"
                ]
            },
...
            "Route": {
                "type": "object",
                "properties": {
                    "profile": {
                        "$ref": "#/components/schemas/RouteProfile"
                    },
                    "distance_mi": {
                        "type": "number",
                        "format": "double"
                    },
                    "planned_driving_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "bounds": {
                        "$ref": "#/components/schemas/Bounds"
                    },
                    "legs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RouteLeg"
                        }
                    }
                },
                "required": [
                    "bounds",
                    "distance_mi",
                    "legs",
                    "planned_driving_h",
                    "profile"
                ]
            },
...
            "Stop": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "kind": {
                        "$ref": "#/components/schemas/StopKind"
                    },
                    "lat": {
                        "type": "number",
                        "format": "double"
                    },
                    "lng": {
                        "type": "number",
                        "format": "double"
                    },
                    "label": {
                        "type": "string"
                    },
                    "label_source": {
                        "$ref": "#/components/schemas/LabelSource"
                    },
                    "arrive_at": {
                        "type": "string"
                    },
                    "depart_at": {
                        "type": "string"
                    },
                    "duration_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "duty_status": {
                        "$ref": "#/components/schemas/DutyStatus"
                    },
                    "cumulative_mi": {
                        "type": "number",
                        "format": "double"
                    },
                    "leg_index": {
                        "type": "integer",
                        "maximum": 1,
                        "minimum": 0
                    },
                    "note": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    }
                },
                "required": [
                    "arrive_at",
                    "cumulative_mi",
                    "depart_at",
                    "duration_h",
                    "duty_status",
                    "id",
                    "kind",
                    "label",
                    "label_source",
                    "lat",
                    "leg_index",
                    "lng",
                    "note",
                    "reason"
                ]
            },
...
            "TimelineEvent": {
                "type": "object",
                "properties": {
                    "start_min": {
                        "type": "integer"
                    },
                    "end_min": {
                        "type": "integer"
                    },
                    "start_at": {
                        "type": "string"
                    },
                    "end_at": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/DutyStatus"
                    },
                    "stop_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "start_label": {
                        "type": "string"
                    },
                    "end_label": {
                        "type": "string"
                    },
                    "start_mi": {
                        "type": "number",
                        "format": "double"
                    },
                    "end_mi": {
                        "type": "number",
                        "format": "double"
                    },
                    "note": {
                        "type": "string"
                    }
                },
                "required": [
                    "end_at",
                    "end_label",
                    "end_mi",
                    "end_min",
                    "note",
                    "start_at",
                    "start_label",
                    "start_mi",
                    "start_min",
                    "status",
                    "stop_id"
                ]
            },
...
            "TripMeta": {
                "type": "object",
                "properties": {
                    "places": {
                        "$ref": "#/components/schemas/ResolvedPlaces"
                    },
                    "timezone": {
                        "$ref": "#/components/schemas/TimezoneInfo"
                    },
                    "start_at": {
                        "type": "string",
                        "description": "ISO 8601 with the frozen trip offset."
                    },
                    "cycle_used_start_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "log_header": {
                        "$ref": "#/components/schemas/LogHeader"
                    },
                    "assumptions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Assumption"
                        }
                    },
                    "warnings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Warning"
                        }
                    }
                },
                "required": [
                    "assumptions",
                    "cycle_used_start_h",
                    "log_header",
                    "places",
                    "start_at",
                    "timezone",
                    "warnings"
                ]
            },
            "TripSummary": {
                "type": "object",
                "properties": {
                    "total_distance_mi": {
                        "type": "number",
                        "format": "double"
                    },
                    "driving_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "on_duty_not_driving_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "on_duty_total_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "trip_duration_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "arrival_at": {
                        "type": "string"
                    },
                    "released_at": {
                        "type": "string"
                    },
                    "sheet_count": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "cycle_used_start_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "cycle_used_end_h": {
                        "type": "number",
                        "format": "double"
                    },
                    "counts": {
                        "$ref": "#/components/schemas/StopCounts"
                    }
                },
                "required": [
                    "arrival_at",
                    "counts",
                    "cycle_used_end_h",
                    "cycle_used_start_h",
                    "driving_h",
                    "on_duty_not_driving_h",
                    "on_duty_total_h",
                    "released_at",
                    "sheet_count",
                    "total_distance_mi",
                    "trip_duration_h"
                ]
            },
...
            "LocationInput": {
                "oneOf": [
                    {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 200
                    },
                    {
                        "$ref": "#/components/schemas/PlaceInput"
                    }
                ]
            }
        }
    }
}